Fixed psql insert rounding to nearest integer problem

+++ Assignment2.java
@@ -409,7 +409,7 @@ public class Assignment2 {
                                 rs.getString("cname").trim() + ":" +
                                 rs.getInt("year") + ":" +
                                 rs.getInt("semester") + ":" +
-                                rs.getInt("grade");
+                                rs.getFloat("grade");
                 }
                 // Close the resultset
                 rs.close();
